[step03][refactor] 구입금액을 int형으로 받도록 수정, PurchasePrice 클래스 필드타입 수정

+++ src/main/java/com/hskim/lotto/model/LottoGame.java
@@ -1,5 +1,6 @@
 package com.hskim.lotto.model;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 public class LottoGame {
@@ -18,7 +19,8 @@ public class LottoGame {
         lottoTickets.getWinTableList(winningTicket)
                 .forEach(winnerStatistics::putData);
 
-        earningRate = new EarningRate(winnerStatistics.getTotalPrizeAmount(), lottoTickets.getTotalTicketPrice());
+        earningRate = new EarningRate(winnerStatistics.getTotalPrizeAmount()
+                , BigDecimal.valueOf(lottoTickets.getTotalTicketPrice()));
     }
 
     public Map<LottoWinTable, Integer> getWinnerMap() {
